Move files from seedu.duke into bookmarked
