+++ src/test/java/bookmarked/BookTest.java
@@ -1,5 +1,6 @@
-package seedu.duke;
+package bookmarked;
 
+import bookmarked.Book;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
